Revamp to not use Preact
