--- interface/htm-preact.js
@@ -1,5 +1,0 @@
-import { h } from "https://esm.sh/preact"
-import htm from "https://esm.sh/htm"
-
-export { html as default }
-export const html = htm.bind(h)
