+++ interface/index.html
@@ -11,9 +11,10 @@
 		<title>Static Template</title>
 
 		<link rel="stylesheet" href="./page.css" />
-		<script type="module" src="./switch.js"></script>
+		<script type="module" src="./fetch.js"></script>
 	</head>
 	<body>
+		<!-- Need to do better than this for <noscript> -->
 		<noscript>
 			<p class="warning">
 				For full functionality of this site you need to <a href="https://enable-javascript.com/">enable JavaScript</a>
